Split integration test suites into focused async modules
Previously, integration tests were bundled into 7 monolithic test files.
Because ExUnit parallelizes test execution across *modules* while
executing tests *serially* within each module, this structure created
two major bottlenecks:

1. Low concurrency at startup: Only 7 worker processes could run at T=0,
   underutilizing runners configured with higher concurrency (e.g. 8
   workers in GitHub Actions).
2. Idle worker tail: Faster modules (such as SQLite or basic generators)
   finished early, leaving most worker processes sitting completely idle
   while the suite's wall-clock time was bounded by the slowest
   monolithic modules (`AppWithScopesTest` and
   `UmbrellaAppWithDefaultsTest`).

This splits the integration test suites into 33 focused, granular
modules all marked with `async: true`:

- Preserve all 53 tests.
- Individual test modules should run under 3m00s.
- Fix a missing `@tag database: :sqlite3` on `test "has a passing test suite (--no-live)"`.
- File/module names follow the pattern
  `app_with_{postgres,mysql,mssql,sqlite3}_adapter{,_auth_{html,live},_html,_json,_live,_scopes}_test.exs`
  (and umbrella counterparts).
- Postgres adapter test modules explicitly named for symmetry with the
  other database adapters (instead of "Default").
- Short, systematic app names per module to guarantee test database
  isolation when executing concurrently against PostgreSQL, MySQL, and
  MSSQL while ensuring all generated code strictly satisfies `mix
  format` line-length limits.
- Update documentation in `integration_test/README.md` with rationale.
