+++ integration_test/test/code_generation/app_with_postgres_adapter_auth_html_test.exs
@@ -1,0 +1,28 @@
+defmodule Phoenix.Integration.CodeGeneration.AppWithPostgresAdapterAuthHtmlTest do
+  use Phoenix.Integration.CodeGeneratorCase, async: true
+
+  describe "phx.gen.auth + bcrypt" do
+    test "has no compilation or formatter warnings (--no-live)" do
+      with_installer_tmp("app_with_postgres_adapter_auth_html", fn tmp_dir ->
+        {app_root_path, _} = generate_phoenix_app(tmp_dir, "pg_auth_html")
+
+        mix_run!(~w(phx.gen.auth Accounts User users --no-live), app_root_path)
+
+        assert_no_compilation_warnings(app_root_path)
+        assert_passes_formatter_check(app_root_path)
+      end)
+    end
+
+    @tag database: :postgresql
+    test "has a passing test suite (--no-live)" do
+      with_installer_tmp("app_with_postgres_adapter_auth_html", fn tmp_dir ->
+        {app_root_path, _} = generate_phoenix_app(tmp_dir, "pg_auth_html")
+
+        mix_run!(~w(phx.gen.auth Accounts User users --no-live), app_root_path)
+
+        drop_test_database(app_root_path)
+        assert_tests_pass(app_root_path)
+      end)
+    end
+  end
+end
